[style] 클래스 패키지명 변경
