+++ src/main/java/stringCalculator/StringCalculator.java
@@ -1,17 +1,45 @@
 package stringCalculator;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.function.BiFunction;
 
 public class StringCalculator {
-    private static final int OPERATOR_INDEX_TERM = 2;
+    private static final List<String> OPERATION_SYMBOL = List.of("+", "-", "*", "/");
+    private static Map<String, BiFunction<Integer, Integer, Integer>> operations = new HashMap<>();
     private static final String BLANK = " ";
     private static final String BLANK_INPUT_MESSAGE = "input value is blank";
     private static final String INVALID_OPERATOR_MESSAGE = "input value is invalid operation symbol";
+
+    static {
+        operations.put("+", StringCalculator::addition);
+        operations.put("-", StringCalculator::subtract);
+        operations.put("*", StringCalculator::multiple);
+        operations.put("/", StringCalculator::division);
+    }
+
+    public static int addition(int value1, int value2) {
+        return value1 + value2;
+    }
+
+    public static int subtract(int value1, int value2) {
+        return value1 - value2;
+    }
+
+    public static int division(int value1, int value2) {
+        return value1 / value2;
+    }
+
+    public static int multiple(int value1, int value2) {
+        return value1 * value2;
+    }
 
     public static int calculation(String input) {
         checkBlank(input);
 
         List<String> tokens = List.of(input.split(BLANK));
+        ckeckOperations(tokens);
 
         return calculate(tokens);
     }
@@ -19,10 +47,10 @@ public class StringCalculator {
     private static int calculate(List<String> tokens) {
         int result = Integer.parseInt(tokens.get(0));
 
-        for (int i = 1; i < tokens.size(); i += OPERATOR_INDEX_TERM) {
+        for (int i = 1; i < tokens.size(); i += 2) {
             String operator = tokens.get(i);
-            int operand = toInt(tokens.get(i + 1));
-            result = Operator.findOperator(operator).apply(result, operand);
+            int operand = Integer.parseInt(tokens.get(i + 1));
+            result = operations.get(operator).apply(result, operand);
         }
         return result;
     }
@@ -33,7 +61,15 @@ public class StringCalculator {
         }
     }
 
-    private static int toInt(String token) {
-        return Integer.parseInt(token);
+    private static void ckeckOperations(List<String> tokens) {
+        for (int i = 1; i < tokens.size(); i += 2) {
+            ckeckOpertaion(tokens.get(i));
+        }
+    }
+
+    private static void ckeckOpertaion(String token) {
+        if (!OPERATION_SYMBOL.contains(token)) {
+            throw new IllegalArgumentException(INVALID_OPERATOR_MESSAGE);
+        }
     }
 }
