+++ src/test/java/stringCalculatorTest/StringCalculatorTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import stringCalculator.StringCalculator;
 
 public class StringCalculatorTest {
     @ParameterizedTest(name = "덧셈만 하는 {0} 의 결과값은 {1} 이다")
